+++ client/src/main/java/typeracer/client/messagehandling/PlayerStateNotificationHandler.java
@@ -1,0 +1,42 @@
+package typeracer.client.messagehandling;
+
+import typeracer.client.ViewController;
+import typeracer.communication.messages.Message;
+import typeracer.communication.messages.server.PlayerStateNotification;
+
+
+/**
+ * Handles PlayerStateNotification messages in a chain of responsibility pattern. If the message is not of
+ * the specified type, it will be passed to the next handler in the chain, if any.
+ */
+public class PlayerStateNotificationHandler implements MessageHandler {
+
+  private final MessageHandler nextHandler;
+  private final ViewController viewController;
+
+  /**
+   * Constructor with the next handler in chain.
+   *
+   * @param nextHandler the next handler in message handling chain
+   */
+  public PlayerStateNotificationHandler(MessageHandler nextHandler, ViewController viewController) {
+    this.nextHandler = nextHandler;
+    this.viewController = viewController;
+  }
+
+  /**
+   * Handles the incoming messages.
+   *
+   * @param message the message to handle
+   */
+  @Override
+  public void handleMessage(Message message) {
+    if (message instanceof PlayerStateNotification playerStateNotification) {
+      viewController.updatePlayerStateInformation(playerStateNotification.getPlayerId(), playerStateNotification.getAccuracy(),
+          playerStateNotification.getProgress(), playerStateNotification.getWpm());
+
+    } else if (nextHandler != null) {
+      nextHandler.handleMessage(message);
+    }
+  }
+}